fix(testing): rename cy 10 migration to 11
