+++ packages/cypress/index.ts
@@ -2,4 +2,4 @@ export { cypressProjectGenerator } from './src/generators/cypress-project/cypres
 export { cypressInitGenerator } from './src/generators/init/init';
 export { conversionGenerator } from './src/generators/convert-tslint-to-eslint/convert-tslint-to-eslint';
 export { cypressComponentProject } from './src/generators/cypress-component-project/cypress-component-project';
-export { migrateCypressProject } from './src/generators/migrate-to-cypress-ten/migrate-to-cypress-ten';
+export { migrateCypressProject } from './src/generators/migrate-to-cypress-11/migrate-to-cypress-11';
